requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
django-environ==0.12.0
django-tenants==3.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
psycopg2==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
